requirements.txt
absl-py==0.9.0
appdirs==1.4.3
appnope==0.1.0
astroid==2.4.2
attrs==19.3.0
backcall==0.1.0
backports.cached-property==1.0.1
black==19.10b0
bleach==3.1.1
blessings==1.7
bpython==0.22.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
clip==1.0
curtsies==0.3.10
cwcwidth==0.1.6
cycler==0.10.0
Cython==0.29.16
dataclasses==0.6
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
ftfy==6.0.3
future==0.18.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
greenlet==1.1.2
grpcio==1.28.1
h5py==2.10.0
html5lib==1.0.1
idna==2.9
imageio==2.8.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab==2.0.0
jupyterlab-server==1.0.6
jupyterthemes==0.20.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lesscpy==0.14.0
llvmlite==0.34.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mne==0.24.1
mpimg==0.0.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.1
nibabel==3.0.1
notebook==6.0.3
numba==0.51.2
numexpr==2.7.1
numpy==1.21.5
oauthlib==3.1.0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pathspec==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycortex==1.2.3
Pygments==2.5.2
pylint==2.6.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.27
pyzmq==19.0.0
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
Shapely==1.8.1.post1
six==1.14.0
statsmodels==0.11.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
terminado==0.8.3
testpath==0.4.4
threadpoolctl==3.1.0
toml==0.10.0
torch==1.7.0
torchextractor==0.3.0
torchgel==1.0.0
torchvision==0.8.1
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
zipp==3.0.0
